action_confirm

+++ models/fulfillement_service_rule.py
@@ -8,7 +8,7 @@ class fulfillement_service_rule(models.Model):
 
     fulfillement_sla_name = fields.Char(string="Name", required=True)
     fulfillement_sla_description = fields.Text(string="Description")
-    fulfillement_sla_value = fields.Integer(string="Pourcentage de remplissage")
+    fulfillement_sla_value = fields.Float(string="Pourcentage de remplissage")
     # is_active = fields.Boolean(default=False, string="Active")
     fulfillement_sla_ids = fields.One2many('res.partner.sla', 'sla_id')
     #partner_ids = fields.Many2many('res.partner', string='Clients')
